server/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
feedparser==6.0.11
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
sgmllib3k==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
xmltodict==0.13.0
